rename: qna-datail component directories name
