+++ src/app/(home)/(posts)/qna/[id]/_components/qna-post/PostingAnswerArea.tsx
@@ -74,7 +74,7 @@ const PostingAnswerArea = ({ content, setContent, setToggleAnswer, setQnaComment
       <div className="flex gap-6 h-12 w-[228px] ml-auto mt-6">
         <button
           type="button"
-          className="text-neutral-500 w-[102px] h-12 bg-neutral-50 rounded-md"
+          className="text-neutral-100 w-[102px] h-12 bg-neutral-50 rounded-md"
           onClick={handleCancleClick}
         >
           취소하기
