+++ src/app/(home)/(posts)/qna/[id]/_components/qna-post/PostingQnaAnswer.tsx
